Save 90 seconds by commenting out optical physics!
Optical physics seems not to be used in the simulation (certainly has no
effect on the values written out), but it does take 90 seconds to
initialize.

This 90s wait is a bit of a mystery, as I use the same optical physics
in my simulations, where I do not observe this delay.